+++ changelog.md
@@ -1,5 +1,6 @@
 # Entire changelog
 - [Entire changelog](#entire-changelog)
+  - [Version 2.4:](#version-24)
   - [Version 2.3.1:](#version-231)
   - [Version 2.3:](#version-23)
   - [Version 2.2.2:](#version-222)
@@ -13,6 +14,14 @@
   - [Version 2.0:](#version-20)
   - [Version 1.0:](#version-10)
 
+## Version 2.4:
+- Added new option `CONFIG`.
+  - This option will let you change the configuration of certain options.
+  - Such as: `Search Again | Show Amount on List | Change Current List | Restore to Defaults`.
+    - **Search Again** *NOT IMPLEMENTED*: If `True` and the program has failed to get the latest chapter of a manga, it'll search again. If `False`, the program will continue with the next manga.
+    - **Show Amount on List** *(int | ALL)*: The amount of mangas the program will show on `SHOW => LIST` option. If `ALL` is selected, the program will show the entire list at once.
+    - **Change Current List** *NOT IMPLEMENTED*: This option will let you change to another `.json` file a.k.a. manga list.
+    - **Restore to Defaults**: This will restore the options above to default values.
 ## Version 2.3.1:
 - Resolved issue [#19](https://github.com/SrAranha/MangaCheck/issues/19).
   - When selecting json file, the input was case-sensitive.